VoicePersona/requirements.txt
absl-py==2.3.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.12.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
configargparse==1.7.1
conformer==0.3.2
contourpy==1.3.0
contractions==0.1.73
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
datasets==4.4.2
decorator==5.2.1
diffusers==0.29.0
dill==0.4.0
distro==1.9.0
deinops==0.8.1
exceptiongroup==1.3.1
face-alignment==1.4.1
fastapi==0.115.6
ffmpy==1.0.0
filelock==3.19.1
flask==2.3.3
flatbuffers==25.12.19
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2024.12.0
gast==0.4.0
gdown==5.2.0
google-auth==2.45.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==4.44.1
gradio-client==1.3.0
grpcio==1.76.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.26.0
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-core==1.3.2
hyperpyyaml==1.2.2
idna==3.11
imageio==2.37.2
importlib-metadata==8.7.0
importlib-resources==6.5.2
inflect==7.0.0
iopath==0.1.10
itsdangerous==2.2.0
jmespath==0.10.0
joblib==1.5.3
kaldifst==1.7.17
keras==2.10.0
keras-preprocessing==1.1.2
kiwisolver==1.4.7
lazy-loader==0.4
libclang==18.1.1
librosa==0.10.2
lightning==2.2.4
lightning-utilities==0.15.2
llvmlite==0.43.0
markdown==3.9
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
modelscope==1.10.0
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
natsort==8.4.0
numba==0.60.0
numpy==1.23.5
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.16.3
openai==1.30.5
openai-whisper==20231117
opencv-contrib-python==4.7.0.72
opt-einsum==3.4.0
orjson==3.11.5
oss2==2.19.1
packaging==24.2
pandas==1.5.2
pillow==10.4.0
pip==25.3
platformdirs==4.4.0
pooch==1.8.2
portalocker==3.2.0
propcache==0.4.1
protobuf==3.19.6
pyahocorasick==2.2.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydub==0.25.1
pygments==2.19.2
pyparsing==3.2.5
pyreadline3==3.5.4
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-speech-features==0.6
pytorch-lightning==2.6.0
pytorch3d==0.7.8
pytz==2025.2
pywin32==311
pyworld==0.3.5
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
resampy==0.4.3
rich==14.2.0
rsa==4.9.1
ruamel-yaml==0.18.17
ruamel-yaml-clib==0.2.15
ruff==0.14.10
safetensors==0.7.0
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.2.1
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.8.1
soxr==1.0.0
speechbrain==1.0.3
starlette==0.41.3
sympy==1.13.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardx==2.6.4
tensorflow-cpu==2.10.0
tensorflow-estimator==2.10.0
tensorflow-intel==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
textsearch==0.0.24
threadpoolctl==3.6.0
tifffile==2024.8.30
tiktoken==0.12.0
tokenizers==0.21.4
tomli==2.3.0
tomlkit==0.12.0
torchaudio==2.5.1
torchmetrics==1.8.2
torchvision==0.20.1
tqdm==4.67.1
transformers==4.51.3
typer==0.20.1
typing-inspection==0.4.2
urllib3==2.6.2
uvicorn==0.30.0
websockets==12.0
werkzeug==3.1.4
wetext==0.1.2
wget==3.2
wrapt==2.0.1
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
